Updated fixtures and created GitHub parser/adapter
Now that we're using templates, it's time to stop checking the output from the events themselves but trust the data that is going into them will be sane and trust that the correct template is called for each one.

The first step was to update the fixtures so they would be more like real ones from GitHub. I also created a little "parser" (not really a parser, more of a wrapper/interpreter I guess) that will turn the giant data into something a little handier for the templates.

The spec still checks the output for now, so I know it still works, but the new thing will be to make sure that data going into the templates is correct (by wrapping the parser with tests). Once we're done that we can stop with all the specific things and just create a template for every case and with only a couple special handlers.